Struct reactive_rs::Broadcast

pub struct Broadcast<'a, C: ?Sized, T: ?Sized> { /* fields omitted */ }

Event source that transmits context/value pairs to multiple observers.

In order to "fork" the broadcast (creating a new stream that will be subscribed to it), the broadcast object can be simply cloned via the Clone trait. Note that cloning the broadcast only increases its reference count; no values are being cloned or copied.

A broadcast may receive a value in one of two ways. First, the user may explicitly call one of its methods: send(), send_ctx(), feed(), feed_ctx(). Second, the broadcast may be created from a parent stream via broadcast() method of the stream object. Either way, each context/value pair received is passed on to each of the subscribed observers, by reference.

Examples

let out = RefCell::new(Vec::new());
let stream = SimpleBroadcast::<i32>::new();
let child1 = stream
    .clone()
    .subscribe(|x| out.borrow_mut().push(*x + 1));
let child2 = stream
    .clone()
    .subscribe(|x| out.borrow_mut().push(*x + 7));
stream.feed(1..=3);
assert_eq!(&*out.borrow(), &[2, 8, 3, 9, 4, 10]);

Methods

impl<'a, C: 'a + ?Sized, T: 'a + ?Sized> Broadcast<'a, C, T>

pub fn new() -> Self

Creates a new broadcast with specified context and item types.

pub fn from_stream<S>(stream: S) -> Self where
S: Stream<'a, Context = C, Item = T>,

Create a broadcast from a stream, enabling multiple observers ("fork" the stream).

Note: this is equivalent to calling broadcast() on the stream object.

pub fn send_ctx<K, B>(&self, ctx: K, value: B) where
K: Borrow<C>,
B: Borrow<T>,

Send a value along with context to all observers of the broadcast.

pub fn send<B>(&self, value: B) where
B: Borrow<T>,
C: Default,

Similar to send_ctx(), but the context is set to the type's default value.

pub fn feed_ctx<K, B, I>(&self, ctx: K, iter: I) where
K: Borrow<C>,
I: Iterator<Item = B>,
B: Borrow<T>,

Convenience method to feed an iterator of values to all observers of the broadcast, along with a given context.

pub fn feed<B, I>(&self, iter: I) where
I: Iterator<Item = B>,
B: Borrow<T>,
C: Default,

Similar to feed_ctx(), but the context is set to the type's default value.